gb: use uint8_t for bitfields
Slight reduction in CPU instructions for ARM Cortex m0+.

+++ peanut_gb.h
@@ -536,10 +536,10 @@ struct gb_s
 
 	struct
 	{
-		unsigned gb_halt	: 1;
-		unsigned gb_ime		: 1;
-		unsigned gb_frame	: 1; /* New frame drawn. */
-		unsigned lcd_blank	: 1;
+		uint8_t gb_halt		: 1;
+		uint8_t gb_ime		: 1;
+		uint8_t gb_frame	: 1; /* New frame drawn. */
+		uint8_t lcd_blank	: 1;
 	};
 
 	/* Cartridge information:
@@ -614,8 +614,8 @@ struct gb_s
 		uint8_t WY;
 
 		/* Only support 30fps frame skip. */
-		unsigned frame_skip_count : 1;
-		unsigned interlace_count : 1;
+		uint8_t frame_skip_count : 1;
+		uint8_t interlace_count : 1;
 	} display;
 
 	/**
@@ -630,21 +630,21 @@ struct gb_s
 		/* Set to enable interlacing. Interlacing will start immediately
 		 * (at the next line drawing).
 		 */
-		unsigned interlace : 1;
-		unsigned frame_skip : 1;
+		uint8_t interlace : 1;
+		uint8_t frame_skip : 1;
 
 		union
 		{
 			struct
 			{
-				unsigned a	: 1;
-				unsigned b	: 1;
-				unsigned select	: 1;
-				unsigned start	: 1;
-				unsigned right	: 1;
-				unsigned left	: 1;
-				unsigned up	: 1;
-				unsigned down	: 1;
+				uint8_t a	: 1;
+				uint8_t b	: 1;
+				uint8_t select	: 1;
+				uint8_t start	: 1;
+				uint8_t right	: 1;
+				uint8_t left	: 1;
+				uint8_t up	: 1;
+				uint8_t down	: 1;
 			} joypad_bits;
 			uint8_t joypad;
 		};
